Fix issues with found counting
Previous code was not correctly accounting for state of previously checked cell. It's necessary to reset `found` to 0 when an empy cell is encountered, so `foundPiece` must reflect when that is the case. Otherwise, non-contiguous lines of like-colored pieces will result in these functions returning true (ending the game).

The additions to the `if` statements preceeding `return true` are also necessary for correct `found` counting. Both pieces colored like previously found pieces, and pieces after an empty space, or on the edge of the board, need to increase the count.